requirements.txt
pyserial>=3.0
matplotlib>=3.0
pandas>=2.0
scipy>=1.10
pytest>=7.0
